Revert back the ordinal transformer to not impute with mediums
Now the ordinal transformer is back to it's original state and sets missing values to a number (usually 0) . This is probably more appropiate for most of the current ordinal categories. Need to seperate out into 2 transformers.
Also reverted the ordinal encoder function to  work with numpy arrays.
This function needs to be made generic for any array type data (see
function for more info).